REFACTOR Change grimin to jolly
It is very confusing to have three trees of namespaces whose root
elements each have a name beginning with "gr". I hadn't thought about
this when I named the Grimoire/Grenada interop namespace root "grimin".
(Which stands for "GRIMoire INput", by the way.)

Now it looks like this interop thing will have to be a library on its
own, making the naming choice even more dangerous. Change grimin to
jolly, because this is a name we've had floating around anyway and it
doesn't cause confusion with anything I know of.