Remove excessive memoization
My original purpose with memoization was that I'm neither (def …)ing
stuff that takes long to load nor do I have to load it everytime I want
to use it. But then I went a bit over the top. Remove memoization on two
functions that are not rerun frequentely on the same arguments and whose
results get memoized by the functions that call them anyway.