jolly.core: Move removing of sentinels
Removing sentinels in the end doesn't work, because
grim-thing->gren-thing already doesn't like them. The only way to keep
grim-t->gren-t self-contained, doing all the reading of extra data, is
to have it return nil on input that it doesn't want to convert. Do it
this way and remove the nils afterwards.